requirements.txt
pytest
pytest-xdist
selenium
allure-pytest
